cover space document navigation flashes
